bug: evaluate_arima_models was failing
The try and except block in the  arima_grid_search.py module was suppressing an error message. The version of ARIMA had changed and it no longer accepted trend=0. I removed this and the effect is that for some pdq combinations it complains because the trend has not been removed. However, not it does work correctly and the RMSE is being computed properly.